Ai4ReleaseNotes-azure-function/requirements.txt
aiohttp
aiosignal
annotated-types
anyio
async-timeout
asyncio
atlassian-python-api
attrs
beautifulsoup4
certifi
charset-normalizer
colorama
dataclasses-json
Deprecated
distro
et-xmlfile
exceptiongroup
frozenlist
greenlet
h11
httpcore
httpx
idna
jmespath
jsonpatch
jsonpointer
langchain
langchain-core
langchain-openai
langchain-text-splitters
langsmith
marshmallow
multidict
mypy-extensions
numpy
oauthlib
openai
openpyxl
orjson
packaging
pip-system-certs
pydantic
python-dotenv
PyYAML
regex
requests
requests-oauthlib
six
sniffio
soupsieve
SQLAlchemy
tenacity
tiktoken
tqdm
typing-inspect
typing_extensions
urllib3
wrapt
yarl
semantic-kernel
# Azure Functions dependencies
azure-functions
